day1/requirements.txt
openai~=1.0
python-dotenv~=1.0
pytest~=8.0
requests~=2.31
